+++ adapter/yandex/src/main/java/org/bidon/yandex/GetYandexTokenUseCase.kt
@@ -1,10 +1,9 @@
 package org.bidon.yandex
 
-import com.yandex.mobile.ads.common.AdType
 import com.yandex.mobile.ads.common.BidderTokenLoadListener
 import com.yandex.mobile.ads.common.BidderTokenLoader
-import com.yandex.mobile.ads.common.BidderTokenRequestConfiguration
-import com.yandex.mobile.ads.common.MobileAds
+import com.yandex.mobile.ads.common.BidderTokenRequest
+import com.yandex.mobile.ads.common.YandexAds
 import kotlinx.coroutines.suspendCancellableCoroutine
 import org.bidon.sdk.BidonSdk
 import org.bidon.sdk.auction.AdTypeParam
@@ -15,45 +14,40 @@ import org.bidon.yandex.ext.toYandexBannerSize
 import kotlin.coroutines.resume
 
 internal class GetYandexTokenUseCase {
-    @Suppress("DEPRECATION")
     suspend operator fun invoke(adTypeParam: AdTypeParam): String? =
         suspendCancellableCoroutine { continuation ->
             val context = adTypeParam.activity.applicationContext
 
             val requestParameters = mapOf(
                 "adapter_network_name" to "Bidon",
-                "adapter_version" to MobileAds.libraryVersion,
+                "adapter_version" to YandexAds.libraryVersion,
                 "adapter_network_sdk_version" to BidonSdk.SdkVersion
             )
 
-            val adType = when (adTypeParam) {
-                is AdTypeParam.Banner -> AdType.BANNER
-                is AdTypeParam.Interstitial -> AdType.INTERSTITIAL
-                is AdTypeParam.Rewarded -> AdType.REWARDED
+            val bidderTokenRequest = when (adTypeParam) {
+                is AdTypeParam.Banner -> {
+                    val bannerAdSize = adTypeParam.bannerFormat.toYandexBannerSize(context)
+                    BidderTokenRequest.banner(bannerAdSize, null, requestParameters)
+                }
+                is AdTypeParam.Interstitial -> {
+                    BidderTokenRequest.interstitial(null, requestParameters)
+                }
+                is AdTypeParam.Rewarded -> {
+                    BidderTokenRequest.rewarded(null, requestParameters)
+                }
             }
 
-            val requestBuilder = BidderTokenRequestConfiguration.Builder(adType)
-
-            if (adTypeParam is AdTypeParam.Banner) {
-                val bannerAdSize = adTypeParam.bannerFormat.toYandexBannerSize(context)
-                requestBuilder.setBannerAdSize(bannerAdSize)
-            }
-
-            val requestConfiguration = requestBuilder
-                .setParameters(requestParameters)
-                .build()
-
-            BidderTokenLoader.loadBidderToken(
-                context = context,
-                bidderTokenRequestConfiguration = requestConfiguration,
+            val loader = BidderTokenLoader(context)
+            loader.loadBidderToken(
+                request = bidderTokenRequest,
                 listener = object : BidderTokenLoadListener {
                     override fun onBidderTokenLoaded(bidderToken: String) {
-                        logInfo(TAG, "Loaded bidder token for $adType")
+                        logInfo(TAG, "Loaded bidder token for ${adTypeParam::class.simpleName}")
                         continuation.resume(bidderToken)
                     }
 
                     override fun onBidderTokenFailedToLoad(failureReason: String) {
-                        logError(TAG, "Error while loading bidder token for $adType: $failureReason", BidonError.NoBid)
+                        logError(TAG, "Error while loading bidder token for ${adTypeParam::class.simpleName}: $failureReason", BidonError.NoBid)
                         continuation.resume(null)
                     }
                 })