+++ adapter/yandex/src/main/java/org/bidon/yandex/impl/YandexInterstitialImpl.kt
@@ -2,7 +2,7 @@ package org.bidon.yandex.impl
 
 import android.app.Activity
 import com.yandex.mobile.ads.common.AdError
-import com.yandex.mobile.ads.common.AdRequestConfiguration
+import com.yandex.mobile.ads.common.AdRequest
 import com.yandex.mobile.ads.common.AdRequestError
 import com.yandex.mobile.ads.common.ImpressionData
 import com.yandex.mobile.ads.interstitial.InterstitialAd
@@ -49,7 +49,7 @@ internal class YandexInterstitialImpl :
         val adUnitId = adParams.adUnitId
             ?: return emitEvent(AdEvent.LoadFailed(BidonError.IncorrectAdUnit(demandId = demandId, message = "adUnitId")))
 
-        val adRequestConfiguration = AdRequestConfiguration.Builder(adUnitId)
+        val adRequest = AdRequest.Builder(adUnitId)
             .apply {
                 when (adParams.adUnit.bidType) {
                     BidType.RTB -> {
@@ -76,7 +76,7 @@ internal class YandexInterstitialImpl :
                 emitEvent(AdEvent.LoadFailed(error.asBidonError()))
             }
         }
-        requestInterstitialAd(adParams.context, adRequestConfiguration, adLoadListener)
+        requestInterstitialAd(adParams.context, adRequest, adLoadListener)
     }
 
     override fun show(activity: Activity) {